client/requirements.txt
websocket-client >= 1.8.0